requirements.txt
asn1crypto==0.23.0
backports-abc==0.5
bcrypt==3.1.4
cffi==1.11.2
coverage==4.4.1
cryptography==2.1.1
Fabric3==1.13.1.post1
idna==2.6
paramiko==2.3.1
pika==0.11.0
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.1.2
redis==2.10.6
six==1.11.0
tornado==4.5.2
